03-Deep-Learning-Projects/01-Computer-Vision-Projects/04-Sign-Language-Recognition/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0 